javaHandsOn/src/test/java/com/example/leetcode/easy/ValidParenthesesTest.java
package com.example.leetcode.easy;

import org.junit.jupiter.api.Assertions;
import org.junit.Test;

public class ValidParenthesesTest {

    @Test
    public void isValidTest(){
        ValidParentheses vp = new ValidParentheses();
        Assertions.assertTrue(() -> vp.isValid("()"));
        Assertions.assertTrue(() -> vp.isValid("()[]{}"));
        Assertions.assertFalse(() -> vp.isValid("("));
        Assertions.assertFalse(() -> vp.isValid("(]"));
        Assertions.assertFalse(() -> vp.isValid("){"));
        Assertions.assertFalse(() -> vp.isValid("[([]])"));
        Assertions.assertFalse(() -> vp.isValid("[({"));
        Assertions.assertTrue(() -> vp.isValid("{[]}"));
        Assertions.assertTrue(() -> vp.isValid("{[]}()"));
    }
}
